Refoator person class
