+++ app.rb
@@ -14,67 +14,24 @@ require './add_person_menu'
 
 require './book_list'
 
+require './manage_person'
+
 # App class
 class App
-  attr_accessor :book_list, :rentals, :persons
+  attr_accessor :book_list, :rentals, :person_list
 
   def initialize
     @book_list = BookList.new
     @rentals = []
-    @persons = []
-  end
-
-
-  def list_persons
-    persons.each { |person| puts "[#{person.class}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}" }
-    puts
-  end
-
-  def create_teacher
-    age = new_positive_number { 'Age: ' }
-    name = new_string { 'Name: ' }
-    specialization = new_string { 'Specialization: ' }
-
-    new_teacher = Teacher.new(specialization, age, name)
-    puts 'Person created successfully' if new_teacher.instance_of?(Teacher)
-    persons << new_teacher
-  end
-
-  def create_student
-    age = new_positive_number { 'Age: ' }
-    name = new_string { 'Name: ' }
-    permission = new_y_n_response { 'Has parent permission? [Y/N]: ' }
-
-    new_student = Student.new(age, name, permission)
-    puts 'Person created successfully' if new_student.instance_of?(Student)
-    persons << new_student
-  end
-
-  def create_person
-    add_person_option = add_person_menu
-    case add_person_option
-    when '1' # Student
-      create_student
-    when '2' # Teacher
-      create_teacher
-    end
-  end
-
-  def list_persons_indexes
-    puts 'Select a person from the following list by number (not id)'
-    index = 0
-    persons.each do |person|
-      puts "#{index}) [#{person.class}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
-      index += 1
-    end
+    @person_list = ManagePerson.new
   end
 
   def create_rental
-    list_books_indexes
-    book = books[new_index(books)]
+    book_list.list_books_indexes
+    book = book_list.return_books_array[new_index(book_list.return_books_array)]
 
-    list_persons_indexes
-    person = persons[new_index(persons)]
+    person_list.list_persons_indexes
+    person = person_list.return_persons_array[new_index(person_list.return_persons_array)]
 
     date = new_string { 'Date: ' }
 
@@ -84,14 +41,8 @@ class App
     rentals << new_rental
   end
 
-  def person_by_id
-    print 'ID of person: '
-    id = user_input
-    persons.each { |person| return person if person.id == id.to_i }
-  end
-
   def list_rentals_by_person_id
-    person = person_by_id
+    person = person_list.person_by_id
 
     return unless person.instance_of?(Student) || person.instance_of?(Teacher)
 
@@ -103,8 +54,8 @@ class App
 
   def books_and_persons_menus(option)
     book_list.print_books if option == 1
-    list_persons if option == 2
-    create_person if option == 3
+    person_list.list_persons if option == 2
+    person_list.create_person if option == 3
     book_list.create_book if option == 4
   end
